+++ R/light_importance.R
@@ -19,22 +19,12 @@
 #' @param metric An optional named list of length one with a metric as element. Defaults to the first metric in the flashlight. The metric needs to be a function with at least four arguments: actual, predicted, case weights w and \code{...}. Irrelevant for \code{type = "shap"}.
 #' @param lower_is_better Logical flag indicating if lower values in the metric are better or not. If set to FALSE, the increase in metric is multiplied by -1. Not used for \code{type = "shap"}.
 #' @param use_linkinv Should retransformation function be applied? Default is FALSE. Not uses for \code{type = "shap"}.
-#' @param metric_name Name of the resulting column containing the name of the metric. Defaults to "metric". Irrelevant for \code{type = "shap"}.
-#' @param value_name Column name in resulting \code{data} containing the variable importance. Defaults to "value".
-#' @param error_name Column name in resulting \code{data} containing the standard error of permutation importance. Defaults to "error".
-#' @param label_name Column name in resulting \code{data} containing the label of the flashlight. Defaults to "label".
-#' @param variable_name Column name in resulting \code{data} containing the variable names. Defaults to "variable".
 #' @param ... Further arguments passed to \code{light_performance}. Not used for \code{type = "shap"}.
 #' @return An object of class \code{light_importance}, \code{light} (and a list) with the following elements.
 #' \itemize{
 #'   \item \code{data} A tibble with results. Can be used to build fully customized visualizations.
 #'   \item \code{by} Same as input \code{by}.
 #'   \item \code{type} Same as input \code{type}. For information only.
-#'   \item \code{metric_name} Column name representing the name of the metric. For information only.
-#'   \item \code{value_name} Same as input \code{value_name}.
-#'   \item \code{error_name} Same as input \code{error_name}.
-#'   \item \code{label_name} Same as input \code{label_name}.
-#'   \item \code{variable_name} Same as input \code{variable_name}.
 #' }
 #' @export
 #' @references Fisher A., Rudin C., Dominici F. (2018). All Models are Wrong but many are Useful: Variable Importance for Black-Box, Proprietary, or Misspecified Prediction Models, using Model Class Reliance. Arxiv.
@@ -57,15 +47,21 @@ light_importance.default <- function(x, ...) {
 #' @export
 light_importance.flashlight <- function(x, data = x$data, by = x$by,
                                         type = c("permutation", "shap"),
-                                        v = NULL, n_max = Inf,
-                                        seed = NULL, m_repetitions = 1,
+                                        v = NULL, n_max = Inf, seed = NULL,
+                                        m_repetitions = 1,
                                         metric = x$metrics[1],
-                                        lower_is_better = TRUE, use_linkinv = FALSE,
-                                        metric_name = "metric",
-                                        value_name = "value", error_name = "error",
-                                        label_name = "label",
-                                        variable_name = "variable", ...) {
+                                        lower_is_better = TRUE,
+                                        use_linkinv = FALSE, ...) {
   type <- match.arg(type)
+
+  warning_on_names(c("metric_name", "value_name", "label_name",
+                     "variable_name", "error_name"), ...)
+
+  metric_name <- getOption("flashlight.metric_name")
+  value_name <- getOption("flashlight.value_name")
+  label_name <- getOption("flashlight.label_name")
+  variable_name <- getOption("flashlight.variable_name")
+  error_name <- getOption("flashlight.error_name")
 
   # Select v; if SHAP, extract data
   if (type == "shap") {
@@ -75,20 +71,26 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
     if (is.null(v)) {
       v <- x$shap$v
     }
-    data <- x$shap$data[x$shap$data[[x$shap$variable_name]] %in% v, ]
+    data <- x$shap$data[x$shap$data[[variable_name]] %in% v, ]
   } else if (is.null(v)) {
     v <- setdiff(colnames(data), c(x$y, by))
   }
-  # Checks
-  key_vars <- c(label_name, metric_name, by)
-  stopifnot(v %in% colnames(data),
-            (n <- nrow(data)) >= 1L,
-            !anyDuplicated(c(key_vars, value_name, variable_name, error_name)))
 
-  # Subsample to n_max
+  # Checks compatible with both shap and permutation importance
+  key_vars <- c(label_name, metric_name, by)
+  stopifnot(
+    "No data!" = is.data.frame(data) && nrow(data) >= 1L,
+    "'by' not in 'data'!" = by %in% colnames(data),
+    "Not all 'v' in 'data'" = v %in% colnames(data),
+    !anyDuplicated(c(key_vars, value_name, variable_name, error_name))
+  )
+  n <- nrow(data)
+
   if (!is.null(seed)) {
     set.seed(seed)
   }
+
+  # Subsample to n_max
   if (n > n_max) {
     data <- data[sample(n, n_max), , drop = FALSE]
   }
@@ -98,11 +100,6 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
     # Calculate variable importance
     data[[value_name]] <- abs(data[["shap_"]])
 
-    # Rename variable column
-    if (x$shap$variable_name != variable_name) {
-      data[[variable_name]] <- data[[x$shap$variable_name]]
-    }
-
     # Group results by variable
     imp <- grouped_stats(data, x = value_name, w = x$w,
                          by = c(by, variable_name), counts = FALSE)
@@ -111,15 +108,20 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
     imp[[label_name]] <- x$label
     imp[ c(error_name, metric_name)] <- NA
   } else {
+    stopifnot(
+      "Need a metric." = !is.null(metric),
+      "Need exactly one metric." = length(metric) == 1L,
+      "No 'y' defined in flashlight!" = !is.null(x$y)
+    )
+
     # Update flashlight with everything except data
-    stopifnot(!is.null(metric), length(metric) == 1L)
     x <- flashlight(x, by = by, metrics = metric,
                     linkinv = if (use_linkinv) x$linkinv else function(z) z)
 
     # Helper function
     perfm <- function(X, vn = "value_original") {
-      light_performance(x, data = X, use_linkinv = TRUE, metric_name = metric_name,
-                        value_name = vn, label_name = label_name, ...)$data
+      rename_one(light_performance(x, data = X, use_linkinv = TRUE, ...)$data,
+                 value_name, vn)
     }
 
     # Performance before shuffling
@@ -136,9 +138,11 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
       mean_error <- function(X) {
         x <- X[["value_shuffled"]]
         x <- x[!is.na(x)]
-        setNames(data.frame(sd(x) / sqrt(length(x)), mean(x)), c(error_name, "value_shuffled"))
+        setNames(data.frame(sd(x) / sqrt(length(x)), mean(x)),
+                 c(error_name, "value_shuffled"))
       }
-      imp <- replicate(m_repetitions, setNames(lapply(v, core_func, S = data), v),
+      imp <- replicate(m_repetitions,
+                       setNames(lapply(v, core_func, S = data), v),
                        simplify = FALSE)
       imp <- unlist(imp, recursive = FALSE)
       imp <- bind_rows(imp, .id = variable_name)
@@ -156,12 +160,8 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
 
   # Organize output
   var_order <- c(key_vars, variable_name, value_name, error_name)
-  out <- list(data = imp[, var_order], by = by,
-              type = type, metric_name = metric_name,
-              value_name = value_name, error_name = error_name,
-              label_name = label_name, variable_name = variable_name)
-  class(out) <- c("light_importance", "light", "list")
-  out
+  add_classes(list(data = imp[, var_order], by = by, type = type),
+              c("light_importance", "light"))
 }
 
 #' @describeIn light_importance Variable importance for a multiflashlight.