+++ R/light_recode.R
@@ -3,7 +3,7 @@
 #' Recodes factor levels of columns in data slots of an object of class \code{light}.
 #'
 #' @param x An object of class \code{light}.
-#' @param what Column identifier in \code{x} (not column name) to be recoded, e.g. "type_name", "label_name".
+#' @param what Column identifier to be recoded, e.g. "type". For backward compatibility, also the option identifier (e.g. "type_name") can be passed.
 #' @param levels Current levels/values of \code{type_name} column (in desired order).
 #' @param labels New levels of \code{type_name} column in same order as \code{levels}.
 #' @param ... Further arguments passed to \code{factor}.
@@ -34,12 +34,14 @@ light_recode.default <- function(x, ...) {
 #' @describeIn light_recode Recoding factors in data slots of \code{light} object.
 #' @export
 light_recode.light <- function(x, what, levels, labels, ...) {
-  stopifnot(what %in% names(x))
-  wt <- x[[what]]
+  if (!is.null(wt <- getOption(paste("flashlight", what, sep = ".")))) {
+    what <- wt
+  }
   data_slots <- names(x)[vapply(x, FUN = is.data.frame, FUN.VALUE = TRUE)]
   for (z in data_slots) {
-    if (wt %in% colnames(x[[z]])) {
-      x[[z]][[wt]] <- factor(x[[z]][[wt]], levels = levels, labels = labels, ...)
+    if (what %in% colnames(x[[z]])) {
+      x[[z]][[what]] <- factor(x[[z]][[what]], levels = levels,
+                               labels = labels, ...)
     }
   }
   x