+++ release/flashlight/R/most_important.R
@@ -4,6 +4,7 @@
 #'
 #' @importFrom utils head
 #' @importFrom dplyr group_by summarize arrange desc across
+#' @importFrom tidyselect all_of
 #' @param x An object of class \code{light_importance}.
 #' @param top_m Maximum number of important variables to be returned. Defaults to \code{Inf}, i.e. return all variables in descending order of importance.
 #' @return A character vector of variable names sorted in descending order by importance.
@@ -28,9 +29,15 @@ most_important.default <- function(x, top_m = Inf) {
 #' @describeIn most_important Extracts most important variables from an object of class \code{light_importance}.
 #' @export
 most_important.light_importance <- function(x, top_m = Inf) {
-  data <- group_by(x$data, across(x$variable_name))
-  total_importance <- summarize(data, across(x$value_name, sum, na.rm = TRUE),
-                                .groups = "drop")
-  total_importance <- arrange(total_importance, across(x$value_name, desc))
-  head(total_importance[[x$variable_name]], top_m)
+  value_name <- getOption("flashlight.value_name")
+  variable_name <- getOption("flashlight.variable_name")
+
+  data <- group_by(x$data, across(all_of(variable_name)))
+  total_importance <- summarize(
+    data, across(all_of(value_name), sum, na.rm = TRUE),
+    .groups = "drop"
+  )
+  total_importance <- arrange(total_importance,
+                              across(all_of(value_name), desc))
+  head(total_importance[[variable_name]], top_m)
 }