+++ release/flashlight/tests/testthat/tests-methods.R
@@ -34,8 +34,9 @@ test_that("light_recode works", {
   fit <- lm(Sepal.Length ~ Species + 0, data = iris)
   fl <- flashlight(model = fit, label = "lm", data = iris, y = "Sepal.Length")
   eff <- light_effects(fl, v = "Species")
-  eff <- light_recode(eff, what = "type_name",
-                      levels = c("response", "predicted", "partial dependence", "ale"),
+  eff <- light_recode(eff, what = "type",
+                      levels = c("response", "predicted",
+                                 "partial dependence", "ale"),
                       labels = c("Observed", "Fitted", "PD", "ALE"))
   expect_equal(as.character(eff$pd$type[1]), "PD")
 })
@@ -90,8 +91,12 @@ test_that("selected 'is' functions work", {
 
   expect_true(is.light(light_breakdown(fl, new_obs = iris[1, ])))
   expect_true(is.light_breakdown(light_breakdown(fl, new_obs = iris[1, ])))
-  expect_true(is.light_breakdown_multi(light_breakdown(fls, new_obs = iris[1, ])))
-  expect_false(is.light_breakdown_multi(light_breakdown(fl, new_obs = iris[1, ])))
+  expect_true(
+    is.light_breakdown_multi(light_breakdown(fls, new_obs = iris[1, ]))
+  )
+  expect_false(
+    is.light_breakdown_multi(light_breakdown(fl, new_obs = iris[1, ]))
+  )
 })
 
 fit <- lm(Sepal.Length ~ Species + 0, data = iris)