+++ tests/testthat/tests-eff.R
@@ -173,7 +173,8 @@ test_that("light_effects works with 'by'", {
 })
 
 test_that("light_scatter works for type response", {
-  sc <- light_scatter(fls, v = "Petal.Length", type = "response", data = iris[1:5, ])
+  sc <- light_scatter(fls, v = "Petal.Length",
+                      type = "response", data = iris[1:5, ])
   expect_equal(sc$data$value, rep(iris$Sepal.Length[1:5], 2))
   expect_true(inherits(plot(sc), "ggplot"))
 })
@@ -189,3 +190,36 @@ test_that("light_scatter works for type residual", {
   expect_equal(mean(sc$data$value), 0)
   expect_true(inherits(plot(sc), "ggplot"))
 })
+
+test_that("Options work for light_scatter", {
+  new_options = list(
+    flashlight.label_name = "ell",
+    flashlight.value_name = "val"
+  )
+  withr::with_options(new_options, {
+    sc <- light_scatter(fl1, v = "Petal.Length", type = "predicted")
+    expect_true(all(c("ell", "val") %in% colnames(sc$data)))
+    expect_true(inherits(plot(sc), "ggplot"))
+  })
+})
+
+test_that("Options work for light_profile", {
+  fit <- lm(Sepal.Length ~ Petal.Width, data = iris)
+  fl <- flashlight(model = fit, label = "lm", data = iris)
+
+  new_options = list(
+    flashlight.label_name = "ell",
+    flashlight.value_name = "val",
+    flashlight.q1_name = "qq1",
+    flashlight.q3_name = "qq3",
+    flashlight.type_name = "tt",
+    flashlight.counts_name = "n"
+  )
+  withr::with_options(new_options, {
+    pd <- light_profile(fl, v = "Petal.Width", stats = "quartiles")
+    expect_true(all(c("ell", "val", "qq1", "qq3", "tt", "n") %in%
+                      colnames(pd$data)))
+    expect_true(inherits(plot(pd), "ggplot"))
+  })
+})
+