+++ NEWS.md
@@ -1,3 +1,13 @@
+# flashlight 0.8.0
+
+## Visible changes
+
+- All `zzz_names` arguments are deprecated and are set as `options(flashlight.zzz_names = ...)` instead.
+
+- `light_ice` and `light_scatter` now use `ggplot`'s `override.aes` to suppress `alpha` in legend.
+
+- Greatly improved error messages.
+
 # flashlight 0.7.5
 
 Another maintainance release to remove dependency on ggpubr.